Renamed test's directory
